INT4 x FP8 fix

+++ include/cutlass/gemm/collective/sm90_mma_array_tma_gmma_rs_warpspecialized_mixed_input.hpp
@@ -871,7 +871,7 @@ public:
       return *reinterpret_cast<float*>(&temp);
     }
     else {
-      return scale;
+      return static_cast<float>(scale);
     }
   }
